transform-neptun-masterdata/requirements.txt
osiris-sdk
apache-beam
configparser
